+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2.1 -- unreleased
 
+- feat: validate provided config
 - fix: auth function asynchronicity
 
 ## 0.2.0 -- 2022-07-25